requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
python-socketio==5.11.4
transformers==4.44.2
torch==2.4.1
sqlalchemy==2.0.36
pydantic==2.9.2
mistralai==1.2.5
requests==2.32.3
python-dotenv==1.0.1
google-generativeai==0.7.2